RF: Move algorithm and subtance input components to  their own files
